[command] Move admin commands into admin package
